requirements.txt
BeautifulSoup
Flask
python-dateutil
requests
sqlalchemy
